+++ .idea/inspectionProfiles/Zone_protect.xml
@@ -139,7 +139,7 @@
       <searchConfiguration name="serialize-no-optional" problemDescriptor="Optional values cannot be saved" text="node($node$).set($value$);" recursive="false" caseInsensitive="false" type="JAVA" pattern_context="default">
         <constraint name="__context__" within="" contains="" />
         <constraint name="node" within="" contains="" />
-        <constraint name="value" target="true" within="" contains="" />
+        <constraint name="value" nameOfExprType="java\.util\.Optional" target="true" within="" contains="" />
       </searchConfiguration>
       <searchConfiguration name="String-Argument-prevent-use" suppressId="allow-string-argument" problemDescriptor="StringArgument can accept anything as valid, this may have issues with other commands, try using another argument where possible" text="new org.zone.commands.system.arguments.simple.StringArgument($id$)" recursive="false" caseInsensitive="false" type="JAVA" pattern_context="default">
         <constraint name="__context__" within="" contains="" />
