+++ src/main/java/org/zone/commands/structure/region/flags/player/damage/falldamage/ZoneFlagPlayerFallDamageEnableDisable.java
@@ -1,4 +1,4 @@
-package org.zone.commands.structure.region.flags.player.entitydamage;
+package org.zone.commands.structure.region.flags.player.damage.falldamage;
 
 import net.kyori.adventure.text.Component;
 import org.jetbrains.annotations.NotNull;
@@ -12,30 +12,31 @@ import org.zone.commands.system.arguments.zone.ZoneArgument;
 import org.zone.commands.system.context.CommandContext;
 import org.zone.region.Zone;
 import org.zone.region.flag.FlagTypes;
-import org.zone.region.flag.player.entitydamage.EntityDamagePlayerFlag;
+import org.zone.region.flag.player.falldamage.PlayerFallDamageFlag;
 import org.zone.utils.Messages;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 
-public class ZoneFlagEntityDamagePlayerEnableDisable implements ArgumentCommand {
+public class ZoneFlagPlayerFallDamageEnableDisable implements ArgumentCommand {
     public static final ExactArgument REGION = new ExactArgument("region");
     public static final ExactArgument FLAG = new ExactArgument("flag");
     public static final ZoneArgument ZONE_VALUE = new ZoneArgument("zone_value",
                                                                    new ZoneArgument.ZoneArgumentPropertiesBuilder());
     public static final ExactArgument PLAYER = new ExactArgument("player");
+    public static final ExactArgument FALL = new ExactArgument("fall");
     public static final ExactArgument DAMAGE = new ExactArgument("damage");
     public static final BooleanArgument ENABLE_DISABLE = new BooleanArgument("enableValue",
-                                                                            "enable", "disable");
+                                                                             "enable", "disable");
     @Override
     public @NotNull List<CommandArgument<?>> getArguments() {
-        return Arrays.asList(REGION, FLAG, ZONE_VALUE, PLAYER, DAMAGE, ENABLE_DISABLE);
+        return Arrays.asList(REGION, FLAG, ZONE_VALUE, PLAYER, FALL, DAMAGE,ENABLE_DISABLE);
     }
 
     @Override
     public @NotNull Component getDescription() {
-        return Component.text("Command to enable/disable the Damage Flag");
+        return Component.text("Command to enable/disable the fall damage flag");
     }
 
     @Override
@@ -47,17 +48,17 @@ public class ZoneFlagEntityDamagePlayerEnableDisable implements ArgumentCommand 
     public @NotNull CommandResult run(@NotNull CommandContext commandContext, @NotNull String... args) {
         boolean enable = commandContext.getArgument(this, ENABLE_DISABLE);
         Zone zone = commandContext.getArgument(this, ZONE_VALUE);
-        EntityDamagePlayerFlag entityDamagePlayerFlag = zone
-                .getFlag(FlagTypes.ENTITY_DAMAGE_PLAYER_FLAG_TYPE)
-                .orElse(new EntityDamagePlayerFlag());
+        PlayerFallDamageFlag playerFallDamageFlag = zone
+                .getFlag(FlagTypes.PLAYER_FALL_DAMAGE_FLAG_TYPE)
+                .orElse(new PlayerFallDamageFlag());
         if (enable) {
-            zone.addFlag(entityDamagePlayerFlag);
+            zone.addFlag(playerFallDamageFlag);
         }else {
-            zone.removeFlag(FlagTypes.ENTITY_DAMAGE_PLAYER_FLAG_TYPE);
+            zone.removeFlag(FlagTypes.PLAYER_FALL_DAMAGE_FLAG_TYPE);
         }
         try {
             zone.save();
-            commandContext.sendMessage(Messages.getUpdatedMessage(FlagTypes.ENTITY_DAMAGE_PLAYER_FLAG_TYPE));
+            commandContext.sendMessage(Messages.getUpdatedMessage(FlagTypes.PLAYER_FALL_DAMAGE_FLAG_TYPE));
         }catch (ConfigurateException ce) {
             ce.printStackTrace();
             return CommandResult.error(Messages.getZoneSavingError(ce));
